requirements.txt
flask
flask-cors
aiohttp
python-dotenv
httpx
asgiref
requests
uvicorn
